requirements.txt
discord.py
asyncio 
python-dotenv 
requests
beautifulsoup4 
pytz
pandas
